LayoutTests/fast/profiler/stop-then-function-call.html

<html>
<head>
<script src="resources/profiler-test-JS-resources.js"></script>
<script>
if (window.layoutTestController)
    layoutTestController.dumpAsText();

var j = 0;
function test(len) {
    for (var i = 0; i < len; ++i)
        ++j;
}

function test2(len) {
    for (var i = 0; i < len; ++i)
        --j;
}

console.profile("Test");
test(1000);
console.profileEnd("Test");
test2(1000);

printProfilesDataWithoutTime();
</script>
</head>

<body>
This page has an anonymous JavaScript function that calls built-in functions.
<br>
<br>
To run this test manually, load it in the browser then load the WebInspector and look at
the profile. There should be two nodes in the profile, a "(program)" node with one child that is "test".
</body>
</html>
